requirements.txt
altgraph==0.17.4
ansicon==1.89.0
anyio==4.9.0
bcrypt==4.3.0
blessed==1.20.0
blinker==1.9.0
bottle==0.13.2
build==1.2.2.post1
CacheControl==0.14.2
cachelib==0.13.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
clr_loader==0.2.7.post0
codefind==0.1.7
colorama==0.4.6
crashtest==0.4.1
cryptography==44.0.2
Deprecated==1.2.18
distlib==0.3.9
dulwich==0.22.8
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
Flask==2.2.5
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-Dance==7.1.0
Flask-Limiter==3.12
Flask-Login==0.6.2
Flask-Mail==0.10.0
Flask-MySQL==1.6.0
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.0.1
freegames==2.5.3
git-filter-repo==2.47.0
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.70.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.6
jinxed==1.3.0
jurigged==0.6.0
keyring==25.6.0
limits==4.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.6.0
msgpack==1.1.0
mysql-connector-python==8.0.29
mysqlclient==2.2.7
oauthlib==3.2.2
ordered-set==4.1.0
ovld==0.4.6
packaging==24.2
pbs-installer==2025.3.17
pefile==2023.2.7
pipenv==2024.4.1
pkginfo==1.12.1.2
platformdirs==4.3.7
poetry==2.1.2
poetry-core==2.1.2
proto-plus==1.26.1
protobuf==3.20.3
proxy_tools==0.1.0
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
PyMySQL==1.1.1
pyparsing==3.2.3
pyproject_hooks==1.2.0
python-dotenv==0.19.2
pythonnet==3.0.5
pywebview==5.4
pywin32-ctypes==0.2.3
RapidFuzz==3.13.0
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9.1
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
tomlkit==0.13.2
trove-classifiers==2025.3.19.19
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
URLObject==2.4.3
virtualenv==20.30.0
waitress==3.0.2
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
zstandard==0.23.0